Proof of Concept works! 🎈
Statuses are calculated and a POST web hook can be added that will be called when the status changes.
I have not yet done extensive testing and not tested how well this works if folders are involved (I hope it will go through them recursively but I have not verified this yet).